Update SmallRye Config to 3.18.1
