+++ bom/application/pom.xml
@@ -47,7 +47,7 @@
         <microprofile-lra.version>2.0.2</microprofile-lra.version>
         <microprofile-openapi.version>4.1.1</microprofile-openapi.version>
         <smallrye-common.version>2.19.0</smallrye-common.version>
-        <smallrye-config.version>3.17.2</smallrye-config.version> <!-- keep in sync with microprofile-config-api -->
+        <smallrye-config.version>3.18.1</smallrye-config.version> <!-- keep in sync with microprofile-config-api -->
         <smallrye-health.version>4.3.0</smallrye-health.version>
         <smallrye-modules.version>1.0.beta1</smallrye-modules.version>
         <smallrye-open-api.version>4.3.3</smallrye-open-api.version>
