+++ devtools/gradle/gradle/libs.versions.toml
@@ -2,7 +2,7 @@
 plugin-publish = "2.1.1"
 
 kotlin = "2.4.0"
-smallrye-config = "3.17.2"
+smallrye-config = "3.18.1"
 json-schema-validator = "2.0.1"
 
 junit = "6.1.2"
